+++ camera/vdi_base/common/adapter/platform/v4l2/src/pipeline_core/nodes/v4l2_source_node/v4l2_source_node.cpp
@@ -100,8 +100,8 @@ RetCode V4L2SourceNode::Start(const int32_t streamId)
 #ifdef V4L2_EMULATOR
         }
 #endif
-        format.fmtdesc.width = wide_;
-        format.fmtdesc.height = high_;
+        format.fmtdesc.width = static_cast<uint32_t>(wide_);
+        format.fmtdesc.height = static_cast<uint32_t>(high_);
         int bufCnt = it->format_.bufferCount_;
         rc = sensorController_->Start(bufCnt, format);
         if (rc == RC_ERROR) {
@@ -179,7 +179,7 @@ void V4L2SourceNode::OnMetadataChanged(const std::shared_ptr<CameraMetadata>& me
         CAMERA_LOGE("meta is nullptr");
         return;
     }
-    constexpr uint32_t DEVICE_STREAM_ID = 0;
+    constexpr int32_t DEVICE_STREAM_ID = 0;
     if (sensorController_ != nullptr) {
         if (GetStreamId(metadata) == DEVICE_STREAM_ID) {
             sensorController_->Configure(metadata);
